clean up base curve and surface and nurbs curve and nurbs surface
